refactor OptimizingType to core
