+++ ams/server/src/main/java/com/netease/arctic/server/process/DefaultOptimizingState.java
@@ -25,6 +25,7 @@ import com.netease.arctic.ams.api.StateField;
 import com.netease.arctic.ams.api.process.OptimizingStage;
 import com.netease.arctic.ams.api.process.OptimizingState;
 import com.netease.arctic.ams.api.process.ProcessStatus;
+import com.netease.arctic.optimizing.OptimizingType;
 import com.netease.arctic.server.persistence.OptimizingStatePersistency;
 import com.netease.arctic.server.persistence.StatedPersistentBase;
 import com.netease.arctic.server.persistence.mapper.ProcessMapper;
