+++ ams/server/src/test/java/com/netease/arctic/server/optimizing/TestMixedHiveOptimizing.java
@@ -21,8 +21,8 @@ package com.netease.arctic.server.optimizing;
 import com.netease.arctic.hive.table.SupportHive;
 import com.netease.arctic.io.ArcticHadoopFileIO;
 import com.netease.arctic.io.MixedDataTestHelpers;
+import com.netease.arctic.optimizing.OptimizingType;
 import com.netease.arctic.server.persistence.OptimizingStatePersistency;
-import com.netease.arctic.server.process.OptimizingType;
 import com.netease.arctic.table.ArcticTable;
 import com.netease.arctic.table.KeyedTable;
 import com.netease.arctic.table.TableProperties;
