+++ ams/server/src/test/java/com/netease/arctic/server/optimizing/flow/checker/FullOptimizingMove2HiveChecker.java
@@ -20,12 +20,12 @@ package com.netease.arctic.server.optimizing.flow.checker;
 
 import com.netease.arctic.optimizing.OptimizingInputProperties;
 import com.netease.arctic.server.optimizing.flow.view.TableDataView;
-import com.netease.arctic.optimizing.OptimizingType;
 import com.netease.arctic.server.process.optimizing.TaskDescriptor;
 import com.netease.arctic.table.ArcticTable;
 import org.apache.commons.collections.CollectionUtils;
 
 import javax.annotation.Nullable;
+
 import java.util.List;
 
 public class FullOptimizingMove2HiveChecker extends AbstractHiveChecker {
