+++ ams/server/src/test/java/com/netease/arctic/server/optimizing/flow/checker/MajorOptimizingChecker.java
@@ -18,7 +18,6 @@
 
 package com.netease.arctic.server.optimizing.flow.checker;
 
-import com.netease.arctic.optimizing.OptimizingType;
 import com.netease.arctic.server.process.optimizing.TaskDescriptor;
 import com.netease.arctic.table.ArcticTable;
 import org.apache.commons.collections.CollectionUtils;
