+++ core/src/main/java/com/netease/arctic/optimizing/plan/PartitionEvaluator.java
@@ -18,7 +18,6 @@
 
 package com.netease.arctic.optimizing.plan;
 
-import com.netease.arctic.optimizing.OptimizingType;
 import org.apache.iceberg.ContentFile;
 import org.apache.iceberg.DataFile;
 import org.apache.iceberg.PartitionSpec;
